Move 'CodableMediaStore' to its own file in production
